Moving static data to services.
Since the backend for this app is not built yet, we use static data for
demoing content. Previously, it used to live inside the components that
used it, which creates confusion and data duplication. This change adds
services for retrieveing all model types that are currently available and
moves the static data to them. Returning the static data will soon be
replaced with calls to the backend.