Removing side nav from Subjects view
This change removes the side nav as a means of navigation between centers
when looking Subjects lists. The previous model was in breach with
Clarity's principle of maximum 2 levels of navigation. Choosing the center
to show subjects for is now done through a select input. This change
enabled some refactoring.