Adding the notion of sections inside for a Visit
This commit introduces the notion of sections for visits. Previously, this
application was making the assumption that each visit is a collection of
metrics that need to be captured. After consulting with experts, it turns
out that a visit is rather a collection of sections. Sections are then
collections of captured metrics.