Adding badges to the trial-subjects view to indicate status
This commit makes the Visits Completed column in the trial-subjects
datagrid a badge. The reason behind this is that badges are good with
representing status. In this case, the badge color represents whether
visits for a given subject have started, are in progress or have all been
completed. Despite of this change, the badge remains clickable and will
still navigate the user to the corresponding Visits view.