Fixing a compile warning
This commit fixes a compiler warning that existed previously. The reason
for it was that we were importing clarity-icons by calling the JS file
directly. After looking up the Clarity documentation, it turned out that
we can replace that by importing 2 of the clarity-icons modules that will
give us access to all icons.