Enabling back MongoDB connection
This commit enables back the code that connects to the MongoDB instance on
server start. It was previously commented out as the database needed some
additional setup and configuration. Currently, this code will try to
connect to a MongoDB that runs on localhost, with default port and is called eir.