Hook backend for admin-users view
This change hooks the previously created backend for the User model to the
admin-users component where this model is used. The user.service uses the
new HttpClientModule that comes with Angular 4.3.x to make HTTP calls to
the Express routes. The component view uses Clarity modals for the User
add, update and delete actions.