Add a login page and implement authentication
This change adds a login page component and does all necessary changes
implement authentication.

The login page is Clarity's login page. All views are protected by a guard
which will redirect to the login page, if the user has not logged in
successfully. Upon filling the login and submitting it, we call the
newly-created authentication.service which makes an HTTP POST request to
the backend at /api/auth/login. This is handled by the auth.route.js,
which is an Express extension. This route validates the parameters of the
request and delegates to the auth.controller.js, which does the actual
authentication.