Fix trial definitions backend data model
This change modifies the manner in which data was represented inside
the backend model classes for trial definitions. Previously, Eir was
using references from MongoDB documents to other documents, whenever
we had to represent a nested relationship (e.g. one object is nested
inside another). With this change, we are moving to a completely flat
structure, where the nested documents do not exist as separate
entities inside the database. This decision makes much more sense,
as the nested documents are only ever used inside the context of their
parent.